Nabd_Backend/Back/requirements.txt
django
djangorestframework
jupyter
pandas
numpy
scipy
matplotlib
scikit-learn

